optimize speed

+++ transformer/src/nn_modules/cuda/Int4llamaDecoder.cu
@@ -4,7 +4,6 @@
 
 #include "Int4llamaDecoder.h"
 #include "utils.h"
-// #include "utils.cuh"
 
 // __global__ void prepare_decoder_attention_mask_float(Matrix3D<float> causal_attention_mask, int length, int past_length) {
 //     int i = blockIdx.x * blockDim.x + threadIdx.x;
@@ -26,7 +25,7 @@ __global__ void prepare_decoder_attention_mask_half(Matrix3D<float16_t> causal_a
     
     if(i < length - past_length && j < length) {
         // float16_t min = __float2half(-FLT_MAX);
-        float16_t min = __float2half(-65504.0f);
+        float16_t min = -65504;
         if (i + past_length < j) {
             causal_attention_mask(0, i, j) = min;
         } else {
@@ -76,6 +75,12 @@ Int4llamaDecoder::Int4llamaDecoder(std::string param_path, const struct model_co
 // Int4llamaDecoder:
 struct Int4llamaDecoder_output Int4llamaDecoder::forward(const struct Int4llamaDecoder_input &input) {
     PROFILE_START(profile_name);
+
+    // cudaEvent_t start, stop;
+    // cudaEventCreate(&start);
+    // cudaEventCreate(&stop);
+    // float milliseconds = 0;
+
     int sqlen = input.input_ids.m_dim_z, past_key_values_length = 0;
     // int batch_size = input.input_ids.m_dim_x;
 
@@ -105,11 +110,15 @@ struct Int4llamaDecoder_output Int4llamaDecoder::forward(const struct Int4llamaD
                    (length + threadsPerBlock.y - 1) / threadsPerBlock.y);
     prepare_decoder_attention_mask_half<<<numBlocks, threadsPerBlock>>>(causal_attention_mask, length, past_length);
 
+
+    // cudaEventRecord(start);
+
     std::vector<Matrix3D<float16_t>> past_keys, past_values;
     for (int i = 0; i < this->layers.size(); i++) {
         if (!input.has_past_keys_values) {
             struct Int4llamaDecoderLayer_input l_i = {hidden_states, causal_attention_mask};
             struct Int4llamaDecoderLayer_output l_o = this->layers[i].forward(l_i);
+
             hidden_states = l_o.hidden_states;
             past_keys.push_back(l_o.past_key_value.first);
             past_values.push_back(l_o.past_key_value.second);
@@ -129,11 +138,20 @@ struct Int4llamaDecoder_output Int4llamaDecoder::forward(const struct Int4llamaD
         }
     }
 
+    // cudaEventRecord(stop);
+    // cudaEventSynchronize(stop);
+    // cudaEventElapsedTime(&milliseconds, start, stop);
+    // printf("this->layers.forward: %.2f ms\n", milliseconds);
+
+
     Matrix3D<float16_t> last_hidden_states(last_hidden_states_buf, 1, sqlen, this->embed_dim);
     this->norm.forward(hidden_states, last_hidden_states);
 
     struct Int4llamaDecoder_output output = {last_hidden_states, past_keys, past_values};
     PROFILE_END(profile_name);
+
+    // cudaEventDestroy(start);
+    // cudaEventDestroy(stop);
 
     return output;
 }
